% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELAccessors.R
\name{.ixNumericToLogical}
\alias{.ixNumericToLogical}
\title{Convert numeric indices to logical}
\usage{
.ixNumericToLogical(idx, len, msg)
}
\arguments{
\item{idx}{a numeric containing indices}

\item{len}{a numeric stating the maximum possible index}

\item{msg}{a character stating the error message to produce when indices are
out of bounds}
}
\value{
a logical containing indices
}
\description{
Convert numeric indices to logical
}
