% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELAccessors.R
\name{.MultipleExperiment.charbound}
\alias{.MultipleExperiment.charbound}
\title{Transform character indices to numeric indices}
\usage{
.MultipleExperiment.charbound(idx, txt, fmt)
}
\arguments{
\item{idx}{a character vector containing character indices}

\item{txt}{a character vector containing the reference list}

\item{fmt}{a character stating the format to use for error reporting}
}
\value{
a numeric vector containing numeric indices
}
\description{
Transform character indices to numeric indices
}
