% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ELAccessors.R, R/ELApply.R,
%   R/ELShow.R
\name{ExperimentList-methods}
\alias{ExperimentList-methods}
\alias{experiments}
\alias{nexp}
\alias{experimentData}
\alias{experimentData<-}
\alias{elapply}
\alias{show}
\alias{colData}
\alias{coerce,ExperimentList,SummarizedExperimentList-method}
\alias{coerce,ExperimentList,RangedSummarizedExperimentList-method}
\alias{coerce,ExperimentList,SingleCellExperimentList-method}
\alias{coerce,ExperimentList,SpatialExperimentList-method}
\alias{experimentNames}
\alias{experimentNames<-}
\alias{experimentData,ExperimentList-method}
\alias{experimentData<-,ExperimentList-method}
\alias{nexp,ExperimentList-method}
\alias{experimentNames,ExperimentList-method}
\alias{experimentNames<-,ExperimentList-method}
\alias{[,SummarizedExperimentList,ANY,ANY,ANY-method}
\alias{[,RangedSummarizedExperimentList,ANY,ANY,ANY-method}
\alias{[,SingleCellExperimentList,ANY,ANY,ANY-method}
\alias{[,SpatialExperimentList,ANY,ANY,ANY-method}
\alias{experiments,ExperimentList-method}
\alias{elapply,ExperimentList-method}
\alias{show,SummarizedExperimentList-method}
\alias{show,RangedSummarizedExperimentList-method}
\alias{show,SingleCellExperimentList-method}
\alias{show,SpatialExperimentList-method}
\title{Methods for ExperimentList objects}
\usage{
nexp(x, ...)

experiments(x, ...)

experimentNames(x, ...)

experimentNames(x, ...) <- value

experimentData(x, ...)

experimentData(x, ...) <- value

elapply(x, ...)

\S4method{experimentData}{ExperimentList}(x)

\S4method{experimentData}{ExperimentList}(x) <- value

\S4method{nexp}{ExperimentList}(x)

\S4method{experimentNames}{ExperimentList}(x)

\S4method{experimentNames}{ExperimentList}(x) <- value

\S4method{[}{SummarizedExperimentList,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{RangedSummarizedExperimentList,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{SingleCellExperimentList,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{SpatialExperimentList,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{experiments}{ExperimentList}(x, change.names = TRUE)

\S4method{elapply}{ExperimentList}(x, FUN, ..., check.names = TRUE, change.names = TRUE)

\S4method{show}{SummarizedExperimentList}(object)

\S4method{show}{RangedSummarizedExperimentList}(object)

\S4method{show}{SingleCellExperimentList}(object)

\S4method{show}{SpatialExperimentList}(object)
}
\arguments{
\item{x}{A \code{\link{SummarizedExperimentList}},
\code{\link{RangedSummarizedExperimentList}},
\code{\link{SingleCellExperimentList}}, or
\code{\link{SpatialExperimentList}} object.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{value}{Replacement value for replacement methods.}

\item{i}{A subscript that can act to subset the rows of \code{x}.}

\item{j}{A subscript that can act to subset the columns of \code{x}.}

\item{exp}{A subscript that can act to subset experiments from \code{x}.}

\item{drop}{A \code{logical(1)}, ignored by these methods.}

\item{change.names}{a logical indicating whether column names of columns in
each individual experiment object should be concatenated with experiment
names to ensure unique column names (default is TRUE).}

\item{FUN}{The function to be applied to each element of \code{x}: see
‘Details’.}

\item{check.names}{a logical specifying whether names of \code{experiments}
should be checked against row names of \code{experimentData} (default is
TRUE).}

\item{object}{A \code{\link{SummarizedExperimentList}},
\code{\link{RangedSummarizedExperimentList}},
\code{\link{SingleCellExperimentList}}, or
\code{\link{SpatialExperimentList}} object.}
}
\value{
Return value varies depending on method, as described below.
}
\description{
The \code{\link{ExperimentList}} class provides a family of
methods to manage and process data from multiple experiments.
}
\details{
Additional details for each type of data attribute are provided
below.
}
\section{Constructor}{
 \describe{ ExperimentList instances are constructed
using the ExperimentList function documented in
?\code{\link{ExperimentList}}. }
}

\section{Accessors}{
 \describe{ In the following code snippets, \code{x} is a
ExperimentList object. \item{\code{experimentData(x), experimentData(x) <-
  value}: }{Get or set the experiment data. \code{value} is a
\code{DataFrame} object. Row names of value must be \code{NULL} or
consistent with the existing experiment names of x.}
\item{\code{colWithExperimentData(x)}: }{Get the column data merged with
experiment data.} \item{\code{nexp(x)}: }{Get the number of experiments in
the \code{ExperimentList} object.} \item{\code{experimentNames(x),
  experimentNames(x) <- value}: }{Get or set the names of experiments.}
\item{\code{experiments(x)}: }{Get a list of the experiments (for example,
a list of \code{SummarizedExperiments}).} }
}

\section{Subsetting}{
 \describe{ In the following code snippets, \code{x} is a
ExperimentList object. \item{\code{x[i, j, exp]}: }{Create or replace a
subset of x. \code{i, j, exp} can be \code{numeric}, \code{logical},
\code{character}, or \code{missing}.} \item{\code{subset(x, subset, select,
  experiment)}: }{Create a subset of \code{x} using an expression subset
referring to columns of \code{rowData(x)} and / or select referring to
column names of \code{colData(x)} and / or experiment referring to column
names of \code{experimentData(x)}.} }
}

\section{Apply}{
 \describe{ In the following code snippets, \code{x} is a
ExperimentList object. \item{\code{elapply(x, FUN, ..., simplify = TRUE,
  check.names = TRUE, change.names = TRUE)}: }{Apply functions to each
experiment within an \code{ExperimentList} and potentially simplify the
results.} }
}

\examples{
example(ExperimentList)

nexp(el)
experimentData(el)
colData(el, experimentData = TRUE)
experimentNames(el)
el[1:5, 1:5]
el[, exp = 'PatientA']
elapply(el, dim)

}
\seealso{
\code{\link{ExperimentList}}, \code{\link{SummarizedExperiment}},
\code{\link{RangedSummarizedExperiment}},
\code{\link{SingleCellExperiment}}, \code{\link{SpatialExperiment}}
}
