% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ELConstructors.R
\name{ExperimentList}
\alias{ExperimentList}
\alias{ExperimentList-class}
\alias{SummarizedExperimentList}
\alias{SummarizedExperimentList-class}
\alias{RangedSummarizedExperimentList}
\alias{RangedSummarizedExperimentList-class}
\alias{SingleCellExperimentList}
\alias{SingleCellExperimentList-class}
\alias{SpatialExperimentList}
\alias{SpatialExperimentList-class}
\title{ExperimentList objects}
\usage{
ExperimentList(
  experiments = SimpleList(),
  experimentData = DataFrame(),
  check.names = TRUE,
  change.names = TRUE
)
}
\arguments{
\item{experiments}{a \code{SimpleList} or \code{list} containing a list of
\code{SummarizedExperiment}, \code{RangedSummarizedExperiment},
\code{SingleCellExperiment}, or \code{SpatialExperiment} objects.}

\item{experimentData}{a \code{DataFrame} or \code{data.frame} containing
experiment specific annotations. The number of rows in
\code{experimentData} should match the number of \code{experiments}.}

\item{check.names}{a logical specifying whether names of \code{experiments}
should be checked against row names of \code{experimentData} (default is
TRUE).}

\item{change.names}{a logical indicating whether column names of columns in
each individual experiment object should be concatenated with experiment
names to ensure unique column names (default is TRUE).}
}
\value{
One of the following objects: \code{SummarizedExperimentList},
\code{RangedSummarizedExperimentList}, \code{SingleCellExperimentList}, or
\code{SpatialExperimentList}.
}
\description{
Classes in \code{ExperimentList} are designed to store lists of
experiments such as \code{SummarizedExperiment},
\code{RangedSummarizedExperiment}, \code{SingleCellExperiment},
\code{SpatialExperiment}. Each experiment object is stored in its own
ExperimentList class respectively (one of \code{SummarizedExperimentList},
\code{RangedSummarizedExperimentList}, \code{SingleCellExperimentList},
\code{SpatialExperimentList}). In addition to storing experiment data, the
class also stores experiment-specific annotations associated with each
experiment, analogous to \code{colData} in a SummarizedExperiment object.
}
\details{
This class is designed to enable analysis of data from multiple
experiments. In the context of transcriptomic data, these could be data
from different studies. For single-cell or spatial transcriptomics data,
these could be data from different samples but from the same study. In
these scenarios, analysis of data from different biological samples is made
possible by \code{ExperimentList} objects.

Similar to the design of \code{SummarizedExperiment}, \code{ExperimentList}
enables storing and subsetting of experiment-specific annotations as part
of the \code{experimentData} stored within each object.
}
\examples{
nrows <- 200; ncols <- 6

#Experiment 1
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
se1 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            colData=colData)

#Experiment 2
counts <- matrix(runif(nrows * ncols, 1, 1e4), nrows)
colData <- DataFrame(Treatment=rep(c("ChIP", "Input"), 3),
                     row.names=LETTERS[1:6])
se2 <- SummarizedExperiment(assays=SimpleList(counts=counts),
                            colData=colData)

#Experiment annotation
experimentData <- DataFrame(age = c(36, 72),
                            sex = c("F", "M"),
                            row.names = c("PatientA", "PatientB"))

#Create ExperimentList
experiments = SimpleList("PatientA" = se1, "PatientB" = se2)
el = ExperimentList(experiments = experiments,
                    experimentData = experimentData)
el
}
