% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ELAccessors.R, R/ELApply.R,
%   R/ELShow.R
\name{MultipleExperiment-methods}
\alias{MultipleExperiment-methods}
\alias{experiments}
\alias{nexp}
\alias{experimentData}
\alias{experimentData<-}
\alias{elapply}
\alias{show}
\alias{colData}
\alias{coerce,MultipleExperiment,SummarizedMultipleExperiment-method}
\alias{coerce,MultipleExperiment,RangedSummarizedMultipleExperiment-method}
\alias{coerce,MultipleExperiment,SingleCellMultipleExperiment-method}
\alias{coerce,MultipleExperiment,SpatialMultipleExperiment-method}
\alias{experimentNames}
\alias{experimentNames<-}
\alias{experimentData,MultipleExperiment-method}
\alias{experimentData<-,MultipleExperiment-method}
\alias{nexp,MultipleExperiment-method}
\alias{experimentNames,MultipleExperiment-method}
\alias{experimentNames<-,MultipleExperiment-method}
\alias{[,SummarizedMultipleExperiment,ANY,ANY,ANY-method}
\alias{[,RangedSummarizedMultipleExperiment,ANY,ANY,ANY-method}
\alias{[,SingleCellMultipleExperiment,ANY,ANY,ANY-method}
\alias{[,SpatialMultipleExperiment,ANY,ANY,ANY-method}
\alias{experiments,MultipleExperiment-method}
\alias{elapply,MultipleExperiment-method}
\alias{show,SummarizedMultipleExperiment-method}
\alias{show,RangedSummarizedMultipleExperiment-method}
\alias{show,SingleCellMultipleExperiment-method}
\alias{show,SpatialMultipleExperiment-method}
\title{Methods for MultipleExperiment objects}
\usage{
nexp(x, ...)

experiments(x, ...)

experimentNames(x, ...)

experimentNames(x, ...) <- value

experimentData(x, ...)

experimentData(x, ...) <- value

elapply(x, ...)

\S4method{experimentData}{MultipleExperiment}(x)

\S4method{experimentData}{MultipleExperiment}(x) <- value

\S4method{nexp}{MultipleExperiment}(x)

\S4method{experimentNames}{MultipleExperiment}(x)

\S4method{experimentNames}{MultipleExperiment}(x) <- value

\S4method{[}{SummarizedMultipleExperiment,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{RangedSummarizedMultipleExperiment,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{SingleCellMultipleExperiment,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{[}{SpatialMultipleExperiment,ANY,ANY,ANY}(x, i, j, ..., exp, drop = TRUE)

\S4method{experiments}{MultipleExperiment}(x, change.names = TRUE)

\S4method{elapply}{MultipleExperiment}(x, FUN, ..., check.names = TRUE, change.names = TRUE)

\S4method{show}{SummarizedMultipleExperiment}(object)

\S4method{show}{RangedSummarizedMultipleExperiment}(object)

\S4method{show}{SingleCellMultipleExperiment}(object)

\S4method{show}{SpatialMultipleExperiment}(object)
}
\arguments{
\item{x}{A \code{\link{SummarizedMultipleExperiment}},
\code{\link{RangedSummarizedMultipleExperiment}},
\code{\link{SingleCellMultipleExperiment}}, or
\code{\link{SpatialMultipleExperiment}} object.}

\item{...}{Further arguments to be passed to or from other methods.}

\item{value}{Replacement value for replacement methods.}

\item{i}{A subscript that can act to subset the rows of \code{x}.}

\item{j}{A subscript that can act to subset the columns of \code{x}.}

\item{exp}{A subscript that can act to subset experiments from \code{x}.}

\item{drop}{A \code{logical(1)}, ignored by these methods.}

\item{change.names}{a logical indicating whether column names of columns in
each individual experiment object should be concatenated with experiment
names to ensure unique column names (default is TRUE).}

\item{FUN}{The function to be applied to each element of \code{x}: see
‘Details’.}

\item{check.names}{a logical specifying whether names of \code{experiments}
should be checked against row names of \code{experimentData} (default is
TRUE).}

\item{object}{A \code{\link{SummarizedMultipleExperiment}},
\code{\link{RangedSummarizedMultipleExperiment}},
\code{\link{SingleCellMultipleExperiment}}, or
\code{\link{SpatialMultipleExperiment}} object.}
}
\value{
Return value varies depending on method, as described below.
}
\description{
The \code{\link{MultipleExperiment}} class provides a family of
methods to manage and process data from multiple experiments.
}
\details{
Additional details for each type of data attribute are provided
below.
}
\section{Constructor}{
 \describe{ MultipleExperiment instances are constructed
using the MultipleExperiment function documented in
?\code{\link{MultipleExperiment}}. }
}

\section{Accessors}{
 \describe{ In the following code snippets, \code{x} is a
MultipleExperiment object. \item{\code{experimentData(x), experimentData(x) <-
  value}: }{Get or set the experiment data. \code{value} is a
\code{DataFrame} object. Row names of value must be \code{NULL} or
consistent with the existing experiment names of x.}
\item{\code{colWithExperimentData(x)}: }{Get the column data merged with
experiment data.} \item{\code{nexp(x)}: }{Get the number of experiments in
the \code{MultipleExperiment} object.} \item{\code{experimentNames(x),
  experimentNames(x) <- value}: }{Get or set the names of experiments.}
\item{\code{experiments(x)}: }{Get a list of the experiments (for example,
a list of \code{SummarizedExperiments}).} }
}

\section{Subsetting}{
 \describe{ In the following code snippets, \code{x} is a
MultipleExperiment object. \item{\code{x[i, j, exp]}: }{Create or replace a
subset of x. \code{i, j, exp} can be \code{numeric}, \code{logical},
\code{character}, or \code{missing}.} \item{\code{subset(x, subset, select,
  experiment)}: }{Create a subset of \code{x} using an expression subset
referring to columns of \code{rowData(x)} and / or select referring to
column names of \code{colData(x)} and / or experiment referring to column
names of \code{experimentData(x)}.} }
}

\section{Apply}{
 \describe{ In the following code snippets, \code{x} is a
MultipleExperiment object. \item{\code{elapply(x, FUN, ..., simplify = TRUE,
  check.names = TRUE, change.names = TRUE)}: }{Apply functions to each
experiment within an \code{MultipleExperiment} and potentially simplify the
results.} }
}

\examples{
example(MultipleExperiment)

nexp(el)
experimentData(el)
colData(el, experimentData = TRUE)
experimentNames(el)
el[1:5, 1:5]
el[, exp = 'PatientA']
elapply(el, dim)

}
\seealso{
\code{\link{MultipleExperiment}}, \code{\link{SummarizedExperiment}},
\code{\link{RangedSummarizedExperiment}},
\code{\link{SingleCellExperiment}}, \code{\link{SpatialExperiment}}
}
