% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{RangedSummarizedExperimentList-class}
\alias{RangedSummarizedExperimentList-class}
\alias{.RangedSummarizedExperimentList}
\title{An S4 class to manage lists of related RangedSummarizedExperiment objects}
\description{
An S4 class to manage lists of related RangedSummarizedExperiment objects
}
\section{Slots}{

\describe{
\item{\code{experimentData}}{A \link{DataFrame} object describing the experiments. Row
names, if present, become the experiment names of the Experiment objects.
The number of rows of the DataFrame must equal the number of experiments.}

\item{\code{experimentIndex}}{A \link{numeric} object containing experiment indices of
experiment each column belongs to.}
}}

